test(shields): share completed recovery fixture
